crm_backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
pydantic==2.10.4
pydantic[email]
python-dotenv>=1.0
